requirements.txt
ape
giza-sdk
brownie
numpy
pprint